refactor: インポート履歴の型とステータスラベルを共通化
AdminDashboard と AdminUnitDetail で重複していた ImportStatus / ImportHistory
型を types/common/import_history.ts に、日本語ラベルを constants/import_status.ts
に集約。色は画面ごとに表現方法が異なるため各画面に残す。
ファイル名は既存の user_role.ts に倣いスネークケース。
