+++ frontend/features/AdminDashboard/types.ts
@@ -1,3 +1,5 @@
+import type { ImportHistory } from "@/types/common/import_history";
+
 export type DashboardStats = {
   student_count: number;
   teacher_count: number;
@@ -5,19 +7,7 @@ export type DashboardStats = {
   total_questions: number;
 };
 
-export type ImportStatus = "pending" | "processing" | "completed" | "failed";
-
-export type RecentImport = {
-  id: number;
-  file_name: string;
-  status: ImportStatus;
-  success_count: number;
-  error_count: number;
-  total_count: number;
-  created_at: string;
-};
-
 export type AdminDashboardData = {
   stats: DashboardStats;
-  recent_imports: RecentImport[];
+  recent_imports: ImportHistory[];
 };
