+++ frontend/features/AdminUnitDetail/types.ts
@@ -1,3 +1,5 @@
+import type { ImportHistory } from "@/types/common/import_history";
+
 export type QuestionChoice = {
   id: number;
   choice_number: number | null;
@@ -23,18 +25,6 @@ export type UnitQuestion = {
   choices: QuestionChoice[];
   hints: QuestionHint[];
   explanations: QuestionExplanation[];
-};
-
-export type ImportStatus = "pending" | "processing" | "completed" | "failed";
-
-export type ImportHistory = {
-  id: number;
-  file_name: string;
-  status: ImportStatus;
-  success_count: number;
-  error_count: number;
-  total_count: number;
-  created_at: string;
 };
 
 export type UnitCourse = {
