+++ frontend/libs/domain/course/courseLabel.test.ts
